+++ src/main/java/com/owlmaddie/datagen/CreatureChatDataGenerator.java
@@ -12,5 +12,6 @@ public class CreatureChatDataGenerator implements DataGeneratorEntrypoint {
         FabricDataGenerator.Pack pack = fabricDataGenerator.createPack();
         pack.addProvider(CreatureChatLootTableProvider::new);
         pack.addProvider(CreatureChatAdvancementProvider::new);
+        pack.addProvider(CreatureChatLangProvider::new);
     }
 }